Review of auth0-users auth0-server and auth0-sandbox expected apis
