+++ lib/errors/ApplicationError.js
@@ -11,7 +11,8 @@ function ApplicationError (code, message, additionalInformation) {
     this.additionalInformation = additionalInformation;
   }
   
-  this.status_code = 500;
+  this.name = 'ApplicationError';
+  this.statusCode = 500;
 }
 
 ApplicationError.prototype = Object.create(AuthorizationError.prototype);
